+++ tests/test_analyse.py
@@ -697,3 +697,48 @@ def test_scale_air_pressure_to_height():
             ref_air_temp_degC=pd.Series([1, 2]),
             ref_height_m=2,
             target_height_m=10)
+
+def test_scale_air_density_to_height():
+    assert bw.scale_air_density_to_height(ref_air_density_kg_m3=1.224, ref_height_m=80, target_height_m=100) == 1.22174
+
+    assert bw.scale_air_density_to_height(ref_air_density_kg_m3=1.224,
+                                          ref_height_m=80, 
+                                          target_height_m=100, 
+                                          lapse_rate_kg_m3_m=-0.0002) == 1.22
+    
+    test_density = bw.calc_air_density(DATA.T2m, DATA.P2m)
+    pd.testing.assert_series_equal(bw.scale_air_density_to_height(
+        ref_air_density_kg_m3=test_density.loc['2016-01-09 17:10':'2016-01-09 18:00'],
+        ref_height_m=2, target_height_m=10), 
+        pd.Series(data = [1.186780, 1.187175, 1.187747, 1.185950, 1.186301 , 1.185686],
+        index = pd.to_datetime(['2016-01-09 17:10:00', '2016-01-09 17:20:00', '2016-01-09 17:30:00',
+                                '2016-01-09 17:40:00', '2016-01-09 17:50:00', '2016-01-09 18:00:00'])),
+                                check_names=False)
+ 
+        
+def test_scale_air_temperature_to_height():
+    assert bw.scale_air_temperature_to_height(10.0065, 10, 11) == 10.0
+
+    assert bw.scale_air_temperature_to_height(ref_air_temperature=10, 
+                                              ref_height_m=12, 
+                                              target_height_m=10,
+                                              lapse_rate_deg_m=-0.001) == 10.002
+    
+    pd.testing.assert_series_equal(bw.scale_air_temperature_to_height(
+        ref_air_temperature=DATA.T2m.loc['2016-01-09 17:10':'2016-01-09 18:00'],
+        ref_height_m=2, 
+        target_height_m=20,
+        lapse_rate_deg_m=-0.001),
+        pd.Series(data=[0.936, 0.845, 0.713, 0.834, 0.753, 0.895], 
+        index = pd.to_datetime(['2016-01-09 17:10:00', '2016-01-09 17:20:00', '2016-01-09 17:30:00',
+                                '2016-01-09 17:40:00', '2016-01-09 17:50:00', '2016-01-09 18:00:00'])),
+                                check_names=False)
+    
+    pd.testing.assert_series_equal(bw.scale_air_temperature_to_height(
+        ref_air_temperature=DATA.T2m.loc['2016-01-09 17:10':'2016-01-09 18:00'],
+        ref_height_m=2, 
+        target_height_m=20),
+        pd.Series(data=[0.837, 0.746, 0.614, 0.735, 0.654, 0.796], 
+        index = pd.to_datetime(['2016-01-09 17:10:00', '2016-01-09 17:20:00', '2016-01-09 17:30:00',
+                                '2016-01-09 17:40:00', '2016-01-09 17:50:00', '2016-01-09 18:00:00'])),
+                                check_names=False)